cos_mama_project_backend-main/src/main/java/com/cosmama/backend/api/model/Event.java
package com.cosmama.backend.api.model;

import java.time.LocalDate;

public class Event {
    private long id;
    private String name;
    private String description;
    private LocalDate startDate;
    private LocalDate endDate;
    private String location;

    public Event(long id, String name, String description, LocalDate endDate, String location) {
        this.id = id;
        this.name = name;
        this.description = description;
        this.startDate = LocalDate.now();
        this.endDate = endDate;
        this.location = location;
    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public LocalDate getEndDate() {
        return endDate;
    }

    public void setEndDate(LocalDate endDate) {
        this.endDate = endDate;
    }

    public String getLocation() {
        return location;
    }

    public void setLocation(String location) {
        this.location = location;
    }
}
